requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.10
cryptography==44.0.0
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.11.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.1.3
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-import-export==4.3.3
django-mathfilters==1.0.0
django-multiselectfield==0.1.13
django-mysql==4.15.0
django-phonenumber-field==8.0.0
django-session-timeout==0.1.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
fonttools==4.55.1
html5lib==1.1
idna==3.10
lxml==5.3.0
more-itertools==10.5.0
numpy==2.1.3
odfpy==1.4.1
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.3
pillow==10.4.0
premailer==3.10.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
pyphen==0.15.0
python-barcode==0.15.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
schedule==1.2.2
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
svglib==1.5.1
tablib==3.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
xhtml2pdf==0.2.16
zipp==3.21.0
zopfli==0.2.3.post1
